backend/shared/requirements.txt
pydantic==2.5.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.1.0
pytest==7.4.3

